+++ .tkb
@@ -1,6 +1,66 @@
 {
   "scope": "Workspace",
   "tasks": {
+    "G-1H47JpSI69kav1qZ-kp": {
+      "id": "G-1H47JpSI69kav1qZ-kp",
+      "description": "Star [this repository](https://github.com/md2docx/md2docx-plugin-template/) for easy access and to show your support",
+      "columnId": "column-todo"
+    },
+    "gbzyhx0FwFfzRoE-iqVeu": {
+      "id": "gbzyhx0FwFfzRoE-iqVeu",
+      "description": "run `yarn rebrand` to rebrand and clean up the repo; the `rebrand` script will automatically rebrand the repo, adjust workflows, and create a commit",
+      "columnId": "column-todo"
+    },
+    "Y5__aEZl3Pbxp7r90HKpV": {
+      "id": "Y5__aEZl3Pbxp7r90HKpV",
+      "description": "Install `pnpm` using `npm i -g pnpm`",
+      "columnId": "column-todo"
+    },
+    "jcpymbi71e702p7n6fwvZ": {
+      "id": "jcpymbi71e702p7n6fwvZ",
+      "description": "Install dependencies using `pnpm`: run `pnpm i`.",
+      "columnId": "column-todo"
+    },
+    "RX4J5v4y5IOe_ucf8pMRT": {
+      "id": "RX4J5v4y5IOe_ucf8pMRT",
+      "description": "🌟 Enable [private vulnerability reporting](https://github.com/md2docx/html/security) (For public repo - do this after updating visibility to public)",
+      "columnId": "column-todo"
+    },
+    "EdPbrbJLllUHfZmCS80f7": {
+      "id": "EdPbrbJLllUHfZmCS80f7",
+      "description": "Set up `CodeCov`\n  - Visit Codecov and set up your repo\n  - Create [repository secret]((https://github.com/md2docx/html/settings/secrets/actions)) for `CODECOV_TOKEN`",
+      "columnId": "column-todo"
+    },
+    "MLLUsAhCKaKxvEXFY0HSq": {
+      "id": "MLLUsAhCKaKxvEXFY0HSq",
+      "description": "Set up `CodeClimate`\n  - Visit CodeClimate and set up your repo\n  - Create [repository secret]((https://github.com/md2docx/html/settings/secrets/actions)) for `CC_TEST_REPORTER_ID`\n  - Add `*.test.*` to ignore patterns on the website\n  - Update Code Climate badge",
+      "columnId": "column-todo"
+    },
+    "gMYfaAh2RABMP8uZRQgNx": {
+      "id": "gMYfaAh2RABMP8uZRQgNx",
+      "description": "Add `NPM_AUTH_TOKEN` to repository secrets to automate package publishing\n  - Log in to your [`npm` account](https://www.npmjs.com/login) and create an automation token\n  - Create a new repository secret `NPM_AUTH_TOKEN`",
+      "columnId": "column-todo"
+    },
+    "_BfuX7quWBANpDK1bI7YM": {
+      "id": "_BfuX7quWBANpDK1bI7YM",
+      "description": "Update description in `lib/package.json`",
+      "columnId": "column-todo"
+    },
+    "dC7QDBLH8BmHUfaYmIt81": {
+      "id": "dC7QDBLH8BmHUfaYmIt81",
+      "description": "(Optional) Add Repo Stats by visiting and setting up [repobeats](https://repobeats.axiom.co/)",
+      "columnId": "column-todo"
+    },
+    "P_NrSJQ8m91Odgz8E1fS6": {
+      "id": "P_NrSJQ8m91Odgz8E1fS6",
+      "description": "Create your library and update examples",
+      "columnId": "column-todo"
+    },
+    "1dRWJhy45E1Rq5wZAmPHt": {
+      "id": "1dRWJhy45E1Rq5wZAmPHt",
+      "description": "Update README as required",
+      "columnId": "column-todo"
+    },
     "55DAZwduwTvlaxaJC8hlX": {
       "id": "55DAZwduwTvlaxaJC8hlX",
       "description": "(Optional) Set up [Deepsource](https://app.deepsource.com/login) for static code analysis",
@@ -47,6 +107,18 @@
       "id": "column-todo",
       "title": "To do",
       "tasksIds": [
+        "G-1H47JpSI69kav1qZ-kp",
+        "Y5__aEZl3Pbxp7r90HKpV",
+        "jcpymbi71e702p7n6fwvZ",
+        "gbzyhx0FwFfzRoE-iqVeu",
+        "RX4J5v4y5IOe_ucf8pMRT",
+        "EdPbrbJLllUHfZmCS80f7",
+        "MLLUsAhCKaKxvEXFY0HSq",
+        "gMYfaAh2RABMP8uZRQgNx",
+        "_BfuX7quWBANpDK1bI7YM",
+        "dC7QDBLH8BmHUfaYmIt81",
+        "P_NrSJQ8m91Odgz8E1fS6",
+        "1dRWJhy45E1Rq5wZAmPHt",
         "55DAZwduwTvlaxaJC8hlX",
         "FDFqCQvLm2mUlnTJ3Gna5",
         "O1qAByMoxhHOR-a_guL91",
