--- lib/CHANGELOG.md
@@ -1,7 +1,0 @@
-# @m2d/emoji
-
-## 0.1.0
-
-### Minor Changes
-
-- b686011: Add options to extend or modify emoji list...
