+++ lib/__tests__/index.test.ts
@@ -4,15 +4,15 @@ import { unified } from "unified";
 import remarkParse from "remark-parse";
 import remarkGfm from "remark-gfm";
 import fs from "fs";
-import { emojiPlugin } from "../src";
+import { htmlPlugin } from "../src";
 
 const markdown = fs.readFileSync("../sample.md", "utf-8");
 
 describe("toDocx", () => {
-  it("should handle emojis", async ({ expect }) => {
+  it("should handle htmls", async ({ expect }) => {
     const mdast = unified().use(remarkParse).use(remarkGfm).parse(markdown);
 
-    const docxBlob = await toDocx(mdast, {}, { plugins: [emojiPlugin()] });
+    const docxBlob = await toDocx(mdast, {}, { plugins: [htmlPlugin()] });
 
     expect(docxBlob).toBeInstanceOf(Blob);
   });
