--- packages/shared/CHANGELOG.md
@@ -1,8 +1,0 @@
-# @repo/shared
-
-## 0.0.1
-
-### Patch Changes
-
-- Updated dependencies [b686011]
-  - @m2d/emoji@0.1.0
