+++ packages/shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@repo/shared",
-  "version": "0.0.1",
+  "version": "0.0.0",
   "private": true,
   "sideEffects": false,
   "main": "./dist/index.js",
@@ -41,7 +41,7 @@
   },
   "dependencies": {
     "@m2d/core": "latest",
-    "@m2d/emoji": "workspace:*",
+    "@m2d/emoji": "latest",
     "@mayank1513/fork-me": "^2.1.3",
     "@repo/scripts": "workspace:*",
     "esbuild-raw-plugin": "^0.2.0",
@@ -56,6 +56,7 @@
     "remark-math": "^6.0.0",
     "remark-parse": "^11.0.0",
     "unified": "^11.0.5",
-    "unist-util-remove-position": "^5.0.0"
+    "unist-util-remove-position": "^5.0.0",
+    "@m2d/html": "workspace:*"
   }
 }
