+++ scripts/rebrand.config.json
@@ -1,6 +1,6 @@
 {
-  "packageName": "@m2d/emoji",
+  "packageName": "@m2d/html",
   "owner": "md2docx",
-  "repo": "md2docx-plugin-template",
-  "title": "Md2docx Plugin Template"
+  "repo": "html",
+  "title": "Html"
 }
